requirements.txt
Django==2.0
djangorestframework==3.8.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
